refactor: extract shared plan-patch handler for tasks, epics, features
Replace copy-pasted PATCH logic across 3 route files with handlePlanPatch().
Behaviour is identical — same fields (plan, planApprovedBy, planApprovedAt,
planProgress for tasks), same 404 check, same audit log on plan update.
