+++ apps/web/src/app/api/tasks/[id]/route.ts
@@ -2,6 +2,7 @@ import { NextRequest, NextResponse } from 'next/server'
 import { prisma } from '@/lib/db'
 import { requireServiceAuth, assertCanModify } from '@/lib/auth'
 import { parseBodyOrError, UpdateTaskSchema } from '@/lib/validate'
+import { handlePlanPatch } from '@/lib/plan-patch'
 
 export async function GET(req: NextRequest, { params }: { params: { id: string } }) {
   await requireServiceAuth(req)
@@ -65,31 +66,7 @@ export async function PUT(req: NextRequest, { params }: { params: { id: string }
 }
 
 export async function PATCH(req: NextRequest, { params }: { params: { id: string } }) {
-  const body = await req.json().catch(() => ({})) as Record<string, unknown>
-  const id = params.id
-
-  const existing = await prisma.task.findUnique({ where: { id } })
-  if (!existing) return NextResponse.json({ error: 'Not found' }, { status: 404 })
-
-  const data: Record<string, unknown> = {}
-  if (body.plan            !== undefined) data.plan            = body.plan
-  if (body.planProgress    !== undefined) data.planProgress    = body.planProgress !== null ? Number(body.planProgress) : null
-  if (body.planApprovedBy  !== undefined) data.planApprovedBy  = body.planApprovedBy
-  if (body.planApprovedAt  !== undefined) data.planApprovedAt  = body.planApprovedAt ? new Date(body.planApprovedAt as string) : null
-
-  if (body.plan !== undefined) {
-    await prisma.auditLog.create({
-      data: {
-        userId: req.headers.get('x-user-id') ?? 'system',
-        action: 'plan.updated',
-        target: `task:${id}`,
-        detail: { field: 'plan', entityType: 'task', entityId: id },
-      },
-    }).catch(() => {})
-  }
-
-  const task = await prisma.task.update({ where: { id }, data })
-  return NextResponse.json(task)
+  return handlePlanPatch('task', params.id, req)
 }
 
 export async function DELETE(req: NextRequest, { params }: { params: { id: string } }) {
